# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: EWAMA Properties Ltd API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /properties:
    get:
      operationId: listProperties
      tags: [properties]
      summary: List all properties
      parameters:
        - in: query
          name: status
          schema:
            type: string
            enum: [available, coming_soon, sold_out, archived]
        - in: query
          name: county
          schema:
            type: string
        - in: query
          name: featured
          schema:
            type: boolean
        - in: query
          name: minPrice
          schema:
            type: number
        - in: query
          name: maxPrice
          schema:
            type: number
        - in: query
          name: plotSize
          schema:
            type: string
        - in: query
          name: search
          schema:
            type: string
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            default: 12
      responses:
        "200":
          description: List of properties
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PropertyListResponse"

    post:
      operationId: createProperty
      tags: [properties]
      summary: Create a new property
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PropertyInput"
      responses:
        "201":
          description: Created property
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Property"
  /properties/{id}:
    get:
      operationId: getProperty
      tags: [properties]
      summary: Get a property by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Property
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Property"
        "404":
          description: Not found

    patch:
      operationId: updateProperty
      tags: [properties]
      summary: Update a property
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PropertyUpdate"
      responses:
        "200":
          description: Updated property
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Property"

    delete:
      operationId: deleteProperty
      tags: [properties]
      summary: Delete a property
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /properties/slug/{slug}:
    get:
      operationId: getPropertyBySlug
      tags: [properties]
      summary: Get a property by slug (public)
      parameters:
        - in: path
          name: slug
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Property
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Property"
        "404":
          description: Not found
  /properties/{id}/duplicate:
    post:
      operationId: duplicateProperty
      tags: [properties]
      summary: Duplicate a property
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "201":
          description: Duplicated property
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Property"
  /properties/{id}/publish:
    post:
      operationId: publishProperty
      tags: [properties]
      summary: Publish a property
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Published
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Property"
  /properties/{id}/archive:
    post:
      operationId: archiveProperty
      tags: [properties]
      summary: Archive a property
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Archived
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Property"
...
  /site-visits:
    get:
      operationId: listSiteVisits
      tags: [siteVisits]
      summary: List site visit bookings
      parameters:
        - in: query
          name: status
          schema:
            type: string
            enum: [pending, approved, rejected, rescheduled, completed]
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Site visits list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SiteVisitListResponse"

    post:
      operationId: createSiteVisit
      tags: [siteVisits]
      summary: Book a site visit (public)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SiteVisitInput"
      responses:
        "201":
          description: Booked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SiteVisit"
  /site-visits/{id}:
    get:
      operationId: getSiteVisit
      tags: [siteVisits]
      summary: Get a site visit
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Site visit
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SiteVisit"

    patch:
      operationId: updateSiteVisit
      tags: [siteVisits]
      summary: Update site visit status
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SiteVisitUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SiteVisit"

  # ─── MEDIA ───────────────────────────────────────────────
...
components:
  schemas:
...
    Property:
      type: object
      required: [id, name, slug, county, location, status, cashPrice, installmentPrice, createdAt, updatedAt]
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        county:
          type: string
        location:
          type: string
        googleMapsLink:
          type: ["string", "null"]
        coordinates:
          type: ["string", "null"]
        heroImage:
          type: ["string", "null"]
        gallery:
          type: array
          items:
            type: string
        shortDescription:
          type: ["string", "null"]
        fullDescription:
          type: ["string", "null"]
        plotSize:
          type: ["string", "null"]
        cashPrice:
          type: number
        installmentPrice:
          type: number
        titleDeedFee:
          type: ["number", "null"]
        availablePhases:
          type: array
          items:
            type: string
        amenities:
          type: array
          items:
            type: string
        investmentHighlights:
          type: array
          items:
            type: string
        nearbyLandmarks:
          type: array
          items:
            type: string
        status:
          type: string
          enum: [draft, available, coming_soon, sold_out, archived]
        featured:
          type: boolean
        seoTitle:
          type: ["string", "null"]
        seoDescription:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    PropertyInput:
      type: object
      required: [name, county, location, cashPrice, installmentPrice]
      properties:
        name:
          type: string
          minLength: 2
        slug:
          type: ["string", "null"]
        county:
          type: string
        location:
          type: string
        googleMapsLink:
          type: ["string", "null"]
        coordinates:
          type: ["string", "null"]
        heroImage:
          type: ["string", "null"]
        gallery:
          type: array
          items:
            type: string
        shortDescription:
          type: ["string", "null"]
        fullDescription:
          type: ["string", "null"]
        plotSize:
          type: ["string", "null"]
        cashPrice:
          type: number
          minimum: 0
        installmentPrice:
          type: number
          minimum: 0
        titleDeedFee:
          type: ["number", "null"]
        availablePhases:
          type: array
          items:
            type: string
        amenities:
          type: array
          items:
            type: string
        investmentHighlights:
          type: array
          items:
            type: string
        nearbyLandmarks:
          type: array
          items:
            type: string
        status:
          type: string
          enum: [draft, available, coming_soon, sold_out]
          default: draft
        featured:
          type: boolean
          default: false
        seoTitle:
          type: ["string", "null"]
        seoDescription:
          type: ["string", "null"]
    PropertyUpdate:
      type: object
      properties:
        name:
          type: string
        slug:
          type: ["string", "null"]
        county:
          type: string
        location:
          type: string
        googleMapsLink:
          type: ["string", "null"]
        coordinates:
          type: ["string", "null"]
        heroImage:
          type: ["string", "null"]
        gallery:
          type: array
          items:
            type: string
        shortDescription:
          type: ["string", "null"]
        fullDescription:
          type: ["string", "null"]
        plotSize:
          type: ["string", "null"]
        cashPrice:
          type: number
        installmentPrice:
          type: number
        titleDeedFee:
          type: ["number", "null"]
        availablePhases:
          type: array
          items:
            type: string
        amenities:
          type: array
          items:
            type: string
        investmentHighlights:
          type: array
          items:
            type: string
        nearbyLandmarks:
          type: array
          items:
            type: string
        status:
          type: string
          enum: [draft, available, coming_soon, sold_out, archived]
        featured:
          type: boolean
        seoTitle:
          type: ["string", "null"]
        seoDescription:
          type: ["string", "null"]
    PropertyListResponse:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Property"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer

    # ─── ARTICLE ───────────────────────────────────────────
...
    SiteVisit:
      type: object
      required: [id, name, email, phone, preferredDate, status, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        propertyId:
          type: ["integer", "null"]
        propertyName:
          type: ["string", "null"]
        preferredDate:
          type: string
          format: date
        preferredTime:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, approved, rejected, rescheduled, completed]
        adminNotes:
          type: ["string", "null"]
        rescheduledDate:
          type: ["string", "null"]
          format: date
        createdAt:
          type: string
          format: date-time
    SiteVisitInput:
      type: object
      required: [name, email, phone, preferredDate]
      properties:
        name:
          type: string
        email:
          type: string

        phone:
          type: string
        propertyId:
          type: ["integer", "null"]
        preferredDate:
          type: string
          format: date
        preferredTime:
          type: ["string", "null"]
    SiteVisitUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [pending, approved, rejected, rescheduled, completed]
        adminNotes:
          type: ["string", "null"]
        rescheduledDate:
          type: ["string", "null"]
          format: date
    SiteVisitListResponse:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/SiteVisit"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer

    # ─── MEDIA ─────────────────────────────────────────────
...