# lib/api-zod/src/generated/types/propertyUpdate.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * EWAMA Properties Ltd API
 * OpenAPI spec version: 0.1.0
 */
import type { PropertyUpdateStatus } from './propertyUpdateStatus';

export interface PropertyUpdate {
  name?: string;
  /** @nullable */
  slug?: string | null;
  county?: string;
  location?: string;
  /** @nullable */
  googleMapsLink?: string | null;
  /** @nullable */
  coordinates?: string | null;
  /** @nullable */
  heroImage?: string | null;
  gallery?: string[];
  /** @nullable */
  shortDescription?: string | null;
  /** @nullable */
  fullDescription?: string | null;
  /** @nullable */
  plotSize?: string | null;
  cashPrice?: number;
  installmentPrice?: number;
  /** @nullable */
  titleDeedFee?: number | null;
  availablePhases?: string[];
  amenities?: string[];
  investmentHighlights?: string[];
  nearbyLandmarks?: string[];
  status?: PropertyUpdateStatus;
  featured?: boolean;
  /** @nullable */
  seoTitle?: string | null;
  /** @nullable */
  seoDescription?: string | null;
}